Cleanup build and add support for User-Agent string.
