+++ .gitignore
@@ -1,29 +1,50 @@
-### Eclipse template
+### JetBrains template
+# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and WebStorm
+# Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
 
-.metadata
-bin/
-tmp/
-*.tmp
-*.bak
-*.swp
-*~.nib
-local.properties
-.settings/
-.loadpath
-.recommenders
+# User-specific stuff
+.idea/
 
-# Eclipse Core
-.project
+# IntelliJ
+out/
 
-# JDT-specific (Eclipse Java Development Tools)
-.classpath
+### Gradle template
+.gradle
+/build/
 
-# Code Recommenders
-.recommenders/
+# Ignore Gradle GUI config
+gradle-app.setting
+
+# Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
+!gradle-wrapper.jar
+
+# Cache of project
+.gradletasknamecache
+
 ### Java template
+# Compiled class file
 *.class
 
-setup.bat
-setup.sh
-temp.txt
-src/examples/BinaryBasic.java
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# build artifacts
+classes/
+
+# Package Files #
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
