+++ .travis.yml
@@ -1,0 +1,6 @@
+language: java
+
+jdk:
+  - oraclejdk8
+  - oraclejdk9
+  - oraclejdk10
