--- keys.json
@@ -1,4 +1,0 @@
-{
-  "id": "your_device_id",
-  "token": "your_device_token"
-}
