--- src/examples/payload.json
@@ -1,16 +1,0 @@
-{
-  "sense": [
-    {
-      "asset": "integer",
-      "value": { "byte": 0, "bytelength": 4, "type": "integer"}
-    },
-    {
-      "asset": "boolean",
-      "value": { "byte": 4, "bytelength": 1, "type": "boolean"}
-    },
-    {
-      "asset": "number",
-      "value": { "byte": 5, "bytelength": 4, "type": "number"}
-    }
-  ]
-}
