+++ src/main/java/io/att/util/Asset.java
@@ -1,5 +1,3 @@
-package io.att.messaging;
-
 /*    _   _ _ _____ _    _              _____     _ _     ___ ___  _  __
  *   /_\ | | |_   _| |_ (_)_ _  __ _ __|_   _|_ _| | |__ / __|   \| |/ /
  *  / _ \| | | | | | ' \| | ' \/ _` (_-< | |/ _` | | / / \__ \ |) | ' <
@@ -21,15 +19,25 @@ package io.att.messaging;
  * limitations under the License.
  */
 
-public class Broker {
+package io.att.util;
 
-  public  static String  broker   = "api.allthingstalk.io";
-  private static boolean ssl      = true;
-  public  static int     port     = ssl==true ? 8883 : 1883;
-  private static String  protocol = ssl==true ? "ssl://" : "tcp://";
-  
-  public static String getUrl()
-  {
-    return protocol + broker + ":" + port;
-  }
+public enum Asset {
+
+    INTEGER("\"integer\""),
+    NUMBER("\"number\""),
+    STRING("\"string\""),
+    BOOLEAN("\"boolean\""),
+    LOCATION("\"object\",\"properties\":{\"latitude\":{\"type\":\"number\"}, \"longitude\":{\"type\":\"number\"}, \"altitude\":{\"type\":\"number\"}}}"),
+    POSITION("\"object\",\"properties\":{\"x\":{\"type\":\"number\"}, \"y\":{\"type\":\"number\"}, \"z\":{\"type\":\"number\"}}}"),
+    COLOR("\"object\",\"properties\":{\"r\":{\"type\":\"integer\",\"minimum\":0,\"maximum\":255},\"g\":{\"type\":\"integer\",\"minimum\":0,\"maximum\":255},\"b\":{\"type\":\"integer\",\"minimum\":0,\"maximum\":255}}}");
+
+    private String profile;
+
+    Asset(String profile) {
+        this.profile = profile;
+    }
+
+    public String getProfile() {
+        return profile;
+    }
 }
