+++ src/main/java/io/att/util/Type.java
@@ -1,7 +1,6 @@
 package io.att.util;
 
 public enum Type {
-
   SENSOR,
   ACTUATOR,
   VIRTUAL;
